@@ -1,80 +1,80 @@
 {
-  "name": "oveleon/contao-component-style-manager",
-  "type": "contao-bundle",
-  "description": "Style and CSS-Class Manager for Contao Open Source CMS",
-  "keywords": [
-    "contao",
-    "styles",
-    "css",
-    "manager"
-  ],
-  "homepage": "https://www.oveleon.de/stylemanager.html",
-  "license": "MIT",
-  "authors": [
-    {
-      "name": "Oveleon",
-      "homepage": "https://oveleon.de/",
-      "role": "Developer"
-    },
-    {
-      "name": "Daniele Sciannimanica",
-      "homepage": "https://github.com/doishub",
-      "role": "Developer"
-    }
-  ],
-  "require": {
-    "php": "^8.1",
-    "ext-json": "*",
-    "ext-dom": "*",
-    "contao/core-bundle": "^4.13 || ^5.3",
-    "doctrine/dbal": "^3.3",
-    "psr/log": "^1.0 || 2.0 || ^3.0",
-    "symfony/config": "^5.4 || ^6.4 || ^7.0",
-    "symfony/console": "^5.4 || ^6.4 || ^7.0",
-    "symfony/dependency-injection": "^5.4 || ^6.4 || ^7.0",
-    "symfony/event-dispatcher": "^5.4 || ^6.4 || ^7.0",
-    "symfony/event-dispatcher-contracts": "^2.0 || ^3.0",
-    "symfony/http-foundation": "^5.4 || ^6.4 || ^7.0",
-    "symfony/http-kernel": "^5.4 || ^6.4 || ^7.0",
-    "symfony/routing": "^5.4 || ^6.4 || ^7.0",
-    "symfony/translation-contracts": "^2.3 || ^3.0",
-    "twig/twig": "^3.8"
-  },
-  "require-dev": {
-    "contao/manager-plugin": "^2.3.1"
-  },
-  "suggest": {
-    "contao-thememanager/core": "A lightweight forward-thinking CSS component-first-framework leveraging flexbox and custom properties"
-  },
-  "conflict": {
-    "contao/core": "*",
-    "contao/manager-plugin": "<2.0 || >=3.0"
-  },
-  "autoload": {
-    "psr-4": {
-      "Oveleon\\ContaoComponentStyleManager\\": "src/"
-    },
-    "classmap": [
-      "contao/"
+    "name": "oveleon/contao-component-style-manager",
+    "type": "contao-bundle",
+    "description": "Style and CSS-Class Manager for Contao Open Source CMS",
+    "keywords": [
+        "contao",
+        "styles",
+        "css",
+        "manager"
     ],
-    "exclude-from-classmap": [
-      "contao/config/",
-      "contao/dca/",
-      "contao/languages/",
-      "contao/templates/"
-    ]
-  },
-  "extra": {
-    "branch-alias": {
-      "dev-main": "3.6.x-dev"
+    "homepage": "https://www.oveleon.de/stylemanager.html",
+    "license": "MIT",
+    "authors": [
+        {
+            "name": "Oveleon",
+            "homepage": "https://oveleon.de/",
+            "role": "Developer"
+        },
+        {
+            "name": "Daniele Sciannimanica",
+            "homepage": "https://github.com/doishub",
+            "role": "Developer"
+        }
+    ],
+    "require": {
+        "php": "^8.2",
+        "ext-json": "*",
+        "ext-dom": "*",
+        "contao/core-bundle": "^5.3",
+        "doctrine/dbal": "^3.3",
+        "psr/log": "^1.1 || 2.0 || ^3.0",
+        "symfony/config": "^6.4 || ^7.0",
+        "symfony/console": "^6.4 || ^7.0",
+        "symfony/dependency-injection": "^6.4 || ^7.2",
+        "symfony/event-dispatcher": "^6.4 || ^7.2",
+        "symfony/event-dispatcher-contracts": "^2.0 || ^3.0",
+        "symfony/http-foundation": "^6.4 || ^7.2",
+        "symfony/http-kernel": "^6.4 || ^7.2",
+        "symfony/routing": "^6.4 || ^7.2",
+        "symfony/translation-contracts": "^3.0",
+        "twig/twig": "^3.8"
     },
-    "contao-manager-plugin": "Oveleon\\ContaoComponentStyleManager\\ContaoManager\\Plugin"
-  },
-  "config": {
-    "allow-plugins": {
-      "contao-components/installer": true,
-      "php-http/discovery": true,
-      "contao/manager-plugin": true
+    "require-dev": {
+        "contao/manager-plugin": "^2.3.1"
+    },
+    "suggest": {
+        "contao-thememanager/core": "A lightweight forward-thinking CSS component-first-framework leveraging flexbox and custom properties"
+    },
+    "conflict": {
+        "contao/core": "*",
+        "contao/manager-plugin": "<2.0 || >=3.0"
+    },
+    "autoload": {
+        "psr-4": {
+            "Oveleon\\ContaoComponentStyleManager\\": "src/"
+        },
+        "classmap": [
+            "contao/"
+        ],
+        "exclude-from-classmap": [
+            "contao/config/",
+            "contao/dca/",
+            "contao/languages/",
+            "contao/templates/"
+        ]
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-main": "3.9.x-dev"
+        },
+        "contao-manager-plugin": "Oveleon\\ContaoComponentStyleManager\\ContaoManager\\Plugin"
+    },
+    "config": {
+        "allow-plugins": {
+            "contao-components/installer": true,
+            "php-http/discovery": true,
+            "contao/manager-plugin": true
+        }
     }
-  }
 }
